Solved a map/reduce problem.
The solution to the first problem was rather easy. I used map and reduce
to turn the array into a total number, which was the multiplication of
the entire array. I then divided this aggregate number with the original
list's entry at each index. This gives an array where the components are
the multiplication of each entry except itself.
The algorithmic complexity is O(2n) so O(n).

In the second solution, I was forbidden from using division. I had to
find the new array without division. I found the answer by multiplying
all entries except the current index, for each index. This is a very
expensive operation.
The algorithmic complexity is about O(n3). This is because of the fact
that we have to run through the lists each iteration.