Solved a nestedParentheses problem
The nested parenthesis problem was lifted from 'The Algorithm Design
Manual' by Steven Skiena.

The solution to the problem lied in the data structure used and how it
correlates to the nature of the problem.

The LIFO nature of the problem renders it as a perfect use case for a
stack data structure. And all that was needed was to design the
algorithm.

The solution was to loop through the string and push to the stack if the
parenthesis was '('; pop from the stack if the parenthesis was ')'. And
then, if the stack was empty and there was still another ')' left, then
it wasn't balanced.